+++ include/sdf/Error.hh
@@ -17,23 +17,68 @@
 #ifndef SDF_ERROR_HH_
 #define SDF_ERROR_HH_
 
+#include <iostream>
 #include <string>
 #include "sdf/system_util.hh"
+
+#ifdef _WIN32
+// Disable warning C4251 which is triggered by
+// std::string
+#pragma warning(push)
+#pragma warning(disable: 4251)
+#endif
 
 namespace sdf
 {
   /// \enum ErrorCode
-  /// \brief Set of error codes.
+  /// \brief Set of error codes. Usually one or more errors are returned in
+  /// an Errors vector. The collection of Errors should be take as a whole,
+  /// where an error toward the beginning of the vector can inform errors
+  /// toward the end of the vector.
+  /// \sa Errors
   enum class ErrorCode
   {
     // \brief No error
     NONE = 0,
 
-    /// \brief Indicates that an SDF attribute is missing.
-    NO_ATTRIBUTE,
-
     /// \brief Indicates that reading an SDF file failed.
-    READ_FILE,
+    FILE_READ,
+
+    /// \brief A duplicate name was found for an element where unique names
+    /// are required.
+    DUPLICATE_NAME,
+
+    /// \brief Indicates that a required SDF attribute is missing.
+    ATTRIBUTE_MISSING,
+
+    /// \brief This error indicates that an SDF attribute is invalid.
+    ATTRIBUTE_INVALID,
+
+    /// \brief This error indicates that an SDF attribute is deprecated.
+    ATTRIBUTE_DEPRECATED,
+
+    /// \brief Indicates that a required SDF element is missing.
+    ELEMENT_MISSING,
+
+    /// \brief This error indicates that an SDF element is invalid.
+    ELEMENT_INVALID,
+
+    /// \brief This error indicates that an SDF element is deprecated.
+    ELEMENT_DEPRECATED,
+
+    /// \brief Indicates that an  incorrect SDF element type was
+    /// encountered. This error is used when an element of certain type is
+    /// expected, and an element of a different type was received.
+    ELEMENT_INCORRECT_TYPE,
+
+    /// \brief A URI is invalid.
+    URI_INVALID,
+
+    /// \brief A error occured while trying to resolve a URI.
+    URI_LOOKUP,
+
+    /// \brief A filesystem directory does not exist.
+    DIRECTORY_NONEXISTANT,
   };
 
   class SDFORMAT_VISIBLE Error
@@ -70,6 +115,19 @@ namespace sdf
     /// \sa explicit operator bool() const
     public: bool operator==(const bool _value) const;
 
+    /// \brief Output operator for an error.
+    /// \param[in,out] _out The output stream.
+    /// \param[in] _err The error to output.
+    /// \return Reference to the given output stream
+    public: friend std::ostream &operator<<(std::ostream &_out,
+                                            const sdf::Error &_err)
+    {
+      _out << "Error Code: "
+        << static_cast<std::underlying_type<sdf::ErrorCode>::type>(_err.Code())
+        << " Msg: " << _err.Message();
+      return _out;
+    }
+
     /// \brief The error code value.
     private: ErrorCode code = ErrorCode::NONE;
 
@@ -86,4 +144,9 @@ namespace sdf
 #endif
   };
 }
+#ifdef _WIN32
+#pragma warning(pop)
+#endif
+
+
 #endif