+++ include/sdf/parser.hh
@@ -49,6 +49,25 @@ namespace sdf
   sdf::SDFPtr readFile(const std::string &_filename);
 
   /// \brief Populate the SDF values from a file
+  /// \param[in] _filename Name of the SDF file
+  /// \param[out] _errors Parsing errors will be appended to this variable.
+  /// \return Populated SDF pointer.
+  SDFORMAT_VISIBLE
+  sdf::SDFPtr readFile(const std::string &_filename, Errors &_errors);
+
+  /// \brief Populate the SDF values from a file
+  ///
+  /// This populates the given sdf pointer from a file. If the file is a URDF
+  /// file it is converted to SDF first. All files are converted to the latest
+  /// SDF version
+  /// \param[in] _filename Name of the SDF file
+  /// \param[in] _sdf Pointer to an SDF object.
+  /// \param[out] _errors Parsing errors will be appended to this variable.
+  /// \return True if successful.
+  SDFORMAT_VISIBLE
+  bool readFile(const std::string &_filename, SDFPtr _sdf, Errors &_errors);
+
+  /// \brief Populate the SDF values from a file
   ///
   /// This populates the given sdf pointer from a file. If the file is a URDF
   /// file it is converted to SDF first. All files are converted to the latest
@@ -64,8 +83,27 @@ namespace sdf
   /// This populates the sdf pointer from a string. If the string is a URDF
   /// string it is converted to SDF first. All string are converted to the
   /// latest SDF version
+  /// \param[out] _errors Parsing errors will be appended to this variable.
+  SDFORMAT_VISIBLE
+  bool readString(const std::string &_xmlString, SDFPtr _sdf, Errors &_errors);
+
+  /// \brief Populate the SDF values from a string
+  ///
+  /// This populates the sdf pointer from a string. If the string is a URDF
+  /// string it is converted to SDF first. All string are converted to the
+  /// latest SDF version
   SDFORMAT_VISIBLE
   bool readString(const std::string &_xmlString, SDFPtr _sdf);
+
+  /// \brief Populate the SDF values from a string
+  ///
+  /// This populates the sdf pointer from a string. If the string is a URDF
+  /// string it is converted to SDF first. All strings are converted to the
+  /// latest SDF version
+  /// \param[out] _errors Parsing errors will be appended to this variable.
+  SDFORMAT_VISIBLE
+  bool readString(const std::string &_xmlString, ElementPtr _sdf,
+      Errors &_errors);
 
   /// \brief Populate the SDF values from a string
   ///